requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.0.0
httpx>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
alembic>=1.17.0
pytest>=9.0.1
pytest-asyncio>=1.3.0
starlette>=0.50.0
authlib
httpx
itsdangerous
pyjwt